TimeStamponEdit.gs
This script is designed to update a timestamp in the target column of the same row whenever a user makes an edit in given range of the "Sheet1" sheet. 

If the edited value is not empty, it sets the current timestamp; otherwise, it clears the timestamp.

This is often used for tracking when certain information was last modified in a spreadsheet.